+++ tiger-test-lib/src/test/java/io/cucumber/core/plugin/report/StepDescriptionTest.java
@@ -21,10 +21,9 @@
 package io.cucumber.core.plugin.report;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
 
-import de.gematik.test.tiger.glue.ResolvableArgument;
+import de.gematik.test.tiger.glue.annotation.ResolvableArgument;
 import io.cucumber.plugin.event.DataTableArgument;
 import io.cucumber.plugin.event.PickleStepTestStep;
 import io.cucumber.plugin.event.Step;
@@ -65,7 +64,7 @@ class StepDescriptionTest {
     when(step.getArgument()).thenReturn(dataTableArgument);
     when(dataTableArgument.cells()).thenReturn(List.of(List.of("1", "2", "3")));
     when(pickleStepTestStep.getCodeLocation()).thenReturn(methodName);
-    assertThat(StepDescription.of(pickleStepTestStep).isMethodResolvable(methodName)).isTrue();
+    assertThat(StepDescription.of(pickleStepTestStep).isMethodResolvable()).isTrue();
   }
 
   @ParameterizedTest
@@ -91,7 +90,7 @@ class StepDescriptionTest {
     when(step.getArgument()).thenReturn(dataTableArgument);
     when(dataTableArgument.cells()).thenReturn(List.of(List.of("1", "2", "3")));
     when(pickleStepTestStep.getCodeLocation()).thenReturn(methodName);
-    assertThat(StepDescription.of(pickleStepTestStep).isMethodResolvable(methodName)).isFalse();
+    assertThat(StepDescription.of(pickleStepTestStep).isMethodResolvable()).isFalse();
   }
 
   @ResolvableArgument